+++ tests/overlays.spec.ts
@@ -1,6 +1,7 @@
 import { test, expect } from '@playwright/test';
+import { hoverOver, hoverOut } from './utils/hoverEvents';
 
-test("contents in the hover-overlay slot work as expected", async ({ page }) => {
+test("contents in the hover-overlay slot work as expected", async ({ page, isMobile }) => {
     await page.goto("/tests/overlays.html");
 
     const hoverVideoPlayer = await page.locator("hover-video-player");
@@ -13,16 +14,16 @@ test("contents in the hover-overlay slot work as expected", async ({ page }) => 
         expect(hoverOverlay).toHaveCSS("transition-delay", "0s, 0.3s"),
     ]);
 
-    await hoverVideoPlayer.hover();
+    await hoverOver(hoverVideoPlayer, isMobile);
 
     await expect(hoverOverlay).toHaveCSS("opacity", "1");
 
-    await page.mouse.move(0, 0);
+    await hoverOut(hoverVideoPlayer, isMobile);
 
     await expect(hoverOverlay).toHaveCSS("opacity", "0");
 });
 
-test("contents in the paused-overlay slot work as expected", async ({ page }) => {
+test("contents in the paused-overlay slot work as expected", async ({ page, isMobile }) => {
     await page.route("**/*.mp4", (route) => new Promise((resolve) => {
         // Add a .25 second delay before resolving the request for the video asset so we can
         // test that the player enters a loading state while waiting for the video to load.
@@ -42,7 +43,7 @@ test("contents in the paused-overlay slot work as expected", async ({ page }) =>
         expect(pausedOverlay).toHaveCSS("transition-delay", "0s, 0s"),
     ])
 
-    await hoverVideoPlayer.hover();
+    await hoverOver(hoverVideoPlayer, isMobile);
 
     await Promise.all([
         expect(hoverVideoPlayer).toHaveAttribute("data-playback-state", "loading"),
@@ -59,7 +60,7 @@ test("contents in the paused-overlay slot work as expected", async ({ page }) =>
         expect(pausedOverlay).toHaveCSS("transition-delay", "0s, 0.5s"),
     ]);
 
-    await page.mouse.move(0, 0);
+    await hoverOut(hoverVideoPlayer, isMobile);
 
     await Promise.all([
         expect(hoverVideoPlayer).toHaveAttribute("data-playback-state", "paused"),
@@ -72,7 +73,7 @@ test("contents in the paused-overlay slot work as expected", async ({ page }) =>
     ]);
 });
 
-test("contents in the loading-overlay slot work as expected", async ({ page }) => {
+test("contents in the loading-overlay slot work as expected", async ({ page, isMobile }) => {
     await page.route("**/*.mp4", (route) => new Promise((resolve) => {
         // Add a 0.5 second delay before resolving the request for the video asset so we can
         // test that the loading overlay fades in while waiting for the video to load.
@@ -96,7 +97,7 @@ test("contents in the loading-overlay slot work as expected", async ({ page }) =
         expect(video).toHaveJSProperty("paused", true),
     ]);
 
-    await hoverVideoPlayer.hover();
+    await hoverOver(hoverVideoPlayer, isMobile);
 
     await Promise.all([
         expect(hoverVideoPlayer).toHaveAttribute("data-playback-state", "loading"),
@@ -118,7 +119,7 @@ test("contents in the loading-overlay slot work as expected", async ({ page }) =
         expect(loadingOverlay).toHaveCSS("transition-delay", "0s, 0.1s"),
     ]);
 
-    await page.mouse.move(0, 0);
+    await hoverOut(hoverVideoPlayer, isMobile);
 
     await Promise.all([
         expect(hoverVideoPlayer).toHaveAttribute("data-playback-state", "paused"),