+++ tests/playbackStartDelay.spec.ts
@@ -1,23 +1,24 @@
 import { test, expect } from '@playwright/test';
+import { hoverOut, hoverOver } from './utils/hoverEvents';
 
-test("playback-start-delay attribute works as expected", async ({ page }) => {
+test("playback-start-delay attribute works as expected", async ({ page, isMobile }) => {
     await page.goto('/tests/playbackStartDelay.html');
 
     const hoverVideoPlayer = await page.locator("hover-video-player");
-    const video = await page.locator("hover-video-player video");
+    const video = await hoverVideoPlayer.locator("video");
 
     await Promise.all([
         expect(hoverVideoPlayer).toHaveAttribute("playback-start-delay", "300"),
         expect(hoverVideoPlayer).toHaveJSProperty("playbackStartDelay", 300),
     ]);
 
-    await hoverVideoPlayer.hover();
+    await hoverOver(hoverVideoPlayer, isMobile);
     await expect(video, "the video should still be paused").toHaveJSProperty("paused", true);
     await page.waitForTimeout(300);
 
     await expect(video, "the video should be playing now").toHaveJSProperty("paused", false);
 
-    await page.mouse.move(0, 0);
+    await hoverOut(hoverVideoPlayer, isMobile);
     await expect(video).toHaveJSProperty("paused", true);
 
     await hoverVideoPlayer.evaluate((_hoverVideoPlayer: HTMLElement & { playbackStartDelay: number }) => {
@@ -26,13 +27,13 @@ test("playback-start-delay attribute works as expected", async ({ page }) => {
 
     await expect(hoverVideoPlayer).toHaveAttribute("playback-start-delay", "500");
 
-    await hoverVideoPlayer.hover();
+    await hoverOver(hoverVideoPlayer, isMobile);
     await expect(video, "the video should still be paused").toHaveJSProperty("paused", true);
     await page.waitForTimeout(500);
 
     await expect(video, "the video should be playing now").toHaveJSProperty("paused", false);
 
-    await page.mouse.move(0, 0);
+    await hoverOut(hoverVideoPlayer, isMobile);
     await expect(video).toHaveJSProperty("paused", true);
 
     await hoverVideoPlayer.evaluate((_hoverVideoPlayer: HTMLElement & { playbackStartDelay: number }) => {
@@ -41,7 +42,6 @@ test("playback-start-delay attribute works as expected", async ({ page }) => {
 
     await expect(hoverVideoPlayer).toHaveJSProperty("playbackStartDelay", 550);
 
-    await page.mouse.move(0, 0);
     await expect(video).toHaveJSProperty("paused", true);
 
     await hoverVideoPlayer.evaluate((_hoverVideoPlayer: HTMLElement & { playbackStartDelay: number }) => {
@@ -60,16 +60,16 @@ test("playback timeouts are canceled if the user mouses away before it can run",
     await page.goto('/tests/playbackStartDelay.html');
 
     const hoverVideoPlayer = await page.locator("hover-video-player");
-    const video = await page.locator("hover-video-player video");
+    const video = await hoverVideoPlayer.locator("video");
 
     await Promise.all([
         expect(hoverVideoPlayer).toHaveAttribute("playback-start-delay", "300"),
         expect(hoverVideoPlayer).toHaveJSProperty("playbackStartDelay", 300),
     ]);
 
-    await hoverVideoPlayer.hover();
+    await hoverOver(hoverVideoPlayer, false);
     await expect(video, "the video should still be paused").toHaveJSProperty("paused", true);
-    await page.mouse.move(0, 0);
+    await hoverOut(hoverVideoPlayer, false);
     await page.waitForTimeout(500);
 
     await expect(video, "the video should still not be playing").toHaveJSProperty("paused", true);